// packages/vscode-extension/src/controlPanel/Types.ts
export type SuiteCloudAuthItem = {
	authId: string;
	companyName: string;
	roleName: string;
};